stock_forecast_project_Dev_version/requirements.txt
pandas 
yfinance 
google-cloud-bigquery 
python-decouple 
stockstats 
numpy